requirements.txt
amqp==5.1.1
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
bandit==1.7.3
billiard==3.6.4.0
black==22.1.0
celery==5.2.6
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.2
django-debug-toolbar==3.2.4
django-filter==21.1
django-redis==5.2.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djangorestframework-stubs==1.4.0
drf-yasg==1.20.0
executing==0.8.3
filelock==3.6.0
flake8==4.0.1
flake8-polyfill==1.0.2
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
identify==2.4.11
idna==3.3
inflection==0.5.1
ipython==8.1.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pbr==5.8.1
pep8-naming==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
pre-commit==2.17.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
stack-data==0.2.0
stevedore==3.5.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
types-pytz==2021.3.5
types-PyYAML==6.0.4
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.13.2
wcwidth==0.2.5
wrapt==1.13.3
